LayoutTests/crypto/subtle/aes-gcm-import-key-encrypt-tagLengths-expected.txt

Test encrypting using AES-GCM with an imported 128bit key and tagLengths

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToHexString(cipherText) is expectedCipherText32
PASS bytesToHexString(cipherText) is expectedCipherText64
PASS bytesToHexString(cipherText) is expectedCipherText96
PASS bytesToHexString(cipherText) is expectedCipherText104
PASS bytesToHexString(cipherText) is expectedCipherText112
PASS bytesToHexString(cipherText) is expectedCipherText120
PASS bytesToHexString(cipherText) is expectedCipherText128
PASS successfullyParsed is true

TEST COMPLETE
